backend/requirements.txt
APScheduler==3.9.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
DateTime==4.4
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
twilio==7.9.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
venmo-api==0.3.1
Werkzeug==2.1.2
zipp==3.8.0
zope.interface==5.4.0
